refactor: Refactor serializers
