+++ packages/@pollyjs/adapter-fetch/src/index.js
@@ -1,5 +1,7 @@
 import Adapter from '@pollyjs/adapter';
-import { Fetch as FetchUtils, RequestBodySerializers } from '@pollyjs/utils';
+import { BrowserSerializers, NodeSerializers } from '@pollyjs/utils';
+
+import serializeHeaders from './utils/serializer-headers';
 
 const { defineProperty } = Object;
 const IS_STUBBED = Symbol();
@@ -31,7 +33,7 @@ export default class FetchAdapter extends Adapter {
       this.handleRequest({
         url,
         method: options.method || 'GET',
-        headers: FetchUtils.serializeHeaders(options.headers),
+        headers: serializeHeaders(options.headers),
         body: options.body,
         requestArguments: [url, options]
       });
@@ -72,7 +74,7 @@ export default class FetchAdapter extends Adapter {
     return this.respond(
       pollyRequest,
       response.status,
-      FetchUtils.serializeHeaders(response.headers),
+      serializeHeaders(response.headers),
       await response.text()
     );
   }
@@ -85,8 +87,8 @@ export default class FetchAdapter extends Adapter {
     const { identifiers } = pollyRequest;
     let body = identifiers.body;
 
-    body = await RequestBodySerializers.Browser.serialize(body);
-    body = RequestBodySerializers.Node.serialize(body);
+    body = await BrowserSerializers.serializeRequestBody(body);
+    body = NodeSerializers.serializeRequestBody(body);
 
     identifiers.body = body;
   }
