+++ packages/@pollyjs/adapter-fetch/src/utils/serializer-headers.js
