+++ packages/@pollyjs/adapter-xhr/src/utils/serialize-response-headers.js
