--- packages/@pollyjs/utils/src/utils/fetch/index.js
@@ -1,3 +1,0 @@
-import serializeHeaders from './serialize-headers';
-
-export default { serializeHeaders };
