--- packages/@pollyjs/utils/src/utils/request-body-serializers/index.js
@@ -1,11 +1,0 @@
-import node from './node';
-import browser from './browser';
-
-export default {
-  Node: {
-    serialize: node
-  },
-  Browser: {
-    serialize: browser
-  }
-};
