+++ packages/@pollyjs/utils/src/utils/serializers/browser/index.js
@@ -1,0 +1,3 @@
+import serializeRequestBody from './request-body';
+
+export default { serializeRequestBody };
