--- packages/@pollyjs/utils/src/utils/xhr/index.js
@@ -1,3 +1,0 @@
-import serializeResponseHeaders from './serialize-response-headers';
-
-export default { serializeResponseHeaders };
